Get rid of deferred layout
